requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
motor==3.7.1
pymongo==4.14.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.1.1
pydantic[email]==2.11.7
email-validator==2.2.0
google-generativeai==0.8.5
razorpay==1.4.1
setuptools>=65.5.1 